requirements.txt
amqp==2.5.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
kombu==4.6.3
Markdown==3.1.1
pytz==2019.2
soupsieve==1.9.2
sqlparse==0.3.0
vine==1.3.0
